OF: Introduce DT overlay support.
Introduce DT overlay support.
Using this functionality it is possible to dynamically overlay a part of
the kernel's tree with another tree that's been dynamically loaded.
It is also possible to remove node and properties.

The creation/destruction of the devices is handled by using the standard
of notification framework with two new notifier actions:
OF_RECONFIG_DYNAMIC_CREATE_DEV, when a new bus device should be created
and OF_RECONFIG_DYNAMIC_DESTROY_DEV, when a bus device must be

Documentation about internal and APIs is provided in
	Documentation/devicetree/overlay-notes.txt

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>