arm: omap: Proper cleanups for omap_device
Removing any omap device always resulted in a crash; turns out
BUS_NOTIFY_DEL_DEVICE is not the last notifier event sent in the
course of removing the device, the correct event is
BUS_NOTIFY_UNBOUND_DRIVER, which still is not the right place to
perform the cleanup. A device callback handles that properly, as
well as making sure the hwmods of the device are shutdown.

Signed-off-by: Pantelis Antoniou <panto@antoniou-consulting.com>