of: Do not free memory at of_node_release option
Introduce an option whether to free or not the memory of a node
and it's properties when of_node_release is called.

The option is located in the chosen node, and its name is
of-node-keep. When this boolean property is missing the default
behaviour of freeing is preserved.

The rational behind this option is as follows:

The life-cycle of nodes and properties does not allow us to release
the memory taken by a device_node. Pointer to properties and nodes
might still be in use in drivers, so any memory free'ing is dangerous.

Simply move all the properties to the deadprops list, and the node
itself to of_alldeadnodes until the life-cycles issues are resolved.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>