fixup: implement of_transaction_entry_revert() using of_transaction_entry_apply()
Apply and revert are identical operations as far as the rest of the
kernel is concerned. They both modify the tree, and the operations
required to modify the tree must proceed in exactly the same order.
Instead of having a completely separate implementation for reverting
changes, use of_transaction_entry_invert() to create a transaction entry
that describes the opposite operation and use
of_transaction_entry_apply() to process it.