fixup: Make sure only one node gets attached at a time
This might not actually make sense, but it fixes a bug in the current
code. A better solution would be to allow multiple nodes to be attached
at once, but that requires some fiddling about with doing the addition
of nodes and the _sysfs finishing work to all of the nodes added. This
means walking through all the descendants.

The plan is to get rid of the of_allnodes pointer entirely, so perhaps
it would be better to defer that change until of_allnodes is killed.