of: dynamic: Handle special properties changes
Some node properties are 'special', that is fields in the node
are pointing at it. These are "name", "device_type", and
the phandles ("phandle", "linux,phandle" & "ibm,handle").

Instead of trying to handle them at node attachment, handle
them when the property changes.

Note that name properties are neither removable or updatable.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>