[RFC] OF: probe order dependency aware of_platform_populate
A nice side-effect of the changes in DTC for supporting overlays
is that it is now possible to do dependency tracking of platform
devices automatically.

This patch implements dependency aware probe order for users
of of_platform_populate.

There are no changes in the syntax of the DTS bindings, the
dependency is generated automatically by the use of phandle
references.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>