of: Portable Device Tree connector
Introducing the portable device tree based connector.

It is a method of describing an expansion connector in such a way
that the hardware definition for devices present on the connector
are portable between any board that has a hardware compatible connector.

The connector will maintain the mapping of resources to
actual SoC level resources.

Take for example a simple four pin connector (FPC) which can
be configured as a serial port, an i2c bus or
general purpose gpios.

In connector terms the connector is defined as follows:

FPC: fpc {
	compatible = "extcon,dt-con";
	status = "okay";

	connector {
		#address-cells = <1>;
		#size-cells = <0>;

		P0: p0 {
			reg = <0>;
			gpio = <&gpio0 10>;
		};

		P1: p1 {
			reg = <1>;
			gpio = <&gpio1 8>;
		};

		P2: p2 {
			reg = <2>;
			gpio = <&gpio0 4>;
		};

		P3: p3 {
			reg = <3>;
			gpio = <&gpio2 0>;
		};
	};

};

The 'reg' property contains the address of the connector pin; it's the
pin# from the schematic usually. If the connector contains more than one
physical connectors the address may be of the form
< connector-nr pin-nr>

The gpio property is present when that pin can be configured as a gpio
and points to the real board level gpio chip and gpio number.

Note that the connector node may contain arbitrary properties that
may be used for informational purposes. I.e. a property may be the SoC
pad name that the connector pin is routed from.

&FPC {
	functions {
		gpio {
			gpio-base = <256>;
		};

		uart {
			params {
				#param-cells = <2>;
				generate-pinctrl;

				txd {
					required;
					connector-pin;
				};

				rxd {
					required;
					connector-pin;
				};

			};

			uart0 {
				device = <&uart0>;
				mux@0 {
					txd = <&P0 &uart0_txd>;
					rxd = <&P1 &uart0_rxd_mux0>,
					      <&P3 &uart0_rxd_mux1>;
				};
			};
			uart1 {
				device = <&uart1>;
				mux@0 {
					txd = <&P2 &uart1_txd>;
					rxd = <&P3 &uart1_rxd>;
				};
			};
		};

		i2c {
			params {
				#param-cells = <2>;
				copy-subdevices;
				generate-pinctrl;

				scl {
					required;
					connector-pin;
				};

				sda {
					required;
					connector-pin;
				};

				clock-frequency {
					copy;
				};
			};
			i2c0 {
				device = <&i2c2>;
				mux@0 {
					scl = <&P1 &i2c2_scl>;
					sda = <&P2 &i2c2_sda>;
				};
			};
		};
	};
};

For each function that the connector provides a functions nodes
containing the possible configurations.

For this connector, all four pins can be GPIOs.
Two uarts are routed to the connector.
uart0 with txd at P0 and rxd at either P1 or P3 via a mux option
uart1 with txd at P2 and rxd at P3.
The i2c2 device is routed with scl at P1 and sda at P2.

The parameters node for each function contains a list of the
supported parameters for this function.

For the uart there are two required parameters txd & rxd and the
are defined as connector-pin type.

For the i2c function the two required parameters are scl & sda
while the clock-frequency property will be copied verbatim to
the real hardware i2c device when instantiated.

&FPC {
	plugged {
		compatible = "simple-bus";
		#address-cells = <1>;
		#size-cells = <0>;

		UART {
			compatible = "dtcon-uart";
			status = "okay";
			txd = <0>;
			rxd = <1>;
		};

		LEDS {
			compatible = "simple-bus";
			#address-cells = <1>;
			#size-cells = <0>;

			GPIO_LEDS: gpio_leds {
				compatible = "dtcon-gpio";
				status = "okay";

				gpio-controller;
				#gpio-cells = <2>;

				pin-list = <2>,
					   <3>;
			};

			LEDS {
				compatible = "gpio-leds";
				status = "okay";

				jp0 {
					label = "jp0";
					gpios = <&GPIO_LEDS 0
						GPIO_ACTIVE_HIGH>;
				};
				jp1 {
					label = "jp1";
					gpios = <&GPIO_LEDS 1
						GPIO_ACTIVE_HIGH>;
				};
			};

			LIGHT_I2C {
				compatible = "dtcon-i2c";
				status = "disabled";

				scl = <1>;
				sda = <2>;

				#address-cells = <1>;
				#size-cells = <0>;

				clock-frequency = <100000>;

				/* Ambient light sensor */
				tsl2550@39 {
					compatible = "tsl,tsl2550";
					reg = <0x39>;
					status = "okay";
				};
			};
		};
	};
};

Any activated device nodes in the plugged node will be instantiated.

This example displays two enabled devices and one disabled.

The UART device selects the uart function at pins #0 & #1. The connector
will locate the backend device (uart0) select the right pinctrl by
injecting pinctrl-0 and pinctrl-names properties that are build
dynamically and activates the uart0 device.

The LEDS simple-bus defines a gpio chip forwarder at the two pins (2 & 3)
and creates a LED device using it.

The LIGHT_I2C device is disabled, but if enabled it will activate the
i2c2 real backend device while copying the clock-frequency property and
the ambient light sensor device.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>

dtcon update