cifs: Error STATUS_INVALID_OWNER should map to EACCES
Error STATUS_INVALID_OWNER maps to -EIO. This was encountered when
a non priviledged user attempts to change ownership on a file to another
user on a share mounted with the cifsacl mount option.

 # chown wintest1 a
chown: changing ownership of ‘a’: Input/output error

The error returned is confusing for users who do not expect this error.
The proposal is to return -EACCES instead when a STATUS_INVALID_OWNER is
returned.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Steve French <smfrench@gmail.com>