list: introduce list_del_poison()
ZONE_DEVICE pages always have an elevated count and will never be on an
lru reclaim list.  That space in 'struct page' can be redirected for other
uses, but for safety introduce a poison value that will always trip
__list_add() to assert.  This allows half of the struct list_head storage
to be reclaimed with some assurance to back up the assumption that the
page count never goes to zero and a list_add() is never attempted.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>